Reworked settings serialization: Now XmlSerializer is used also for CLR types.

+++ MP-II/Source/System/MediaPortal.Core/Services/Settings/SettingsManager.cs
@@ -1,4 +1,4 @@
-﻿#region Copyright (C) 2007-2008 Team MediaPortal
+#region Copyright (C) 2007-2008 Team MediaPortal
 
 /*
     Copyright (C) 2007-2008 Team MediaPortal
@@ -24,7 +24,9 @@
 
 using System;
 using System.IO;
+using System.Reflection;
 using MediaPortal.Core;
+using MediaPortal.Core.Logging;
 using MediaPortal.Core.PathManager;
 using MediaPortal.Core.Settings;
 
@@ -55,32 +57,121 @@ namespace MediaPortal.Core.Services.Settings
     #region Protected methods
 
     /// <summary>
-    /// Returns the full file path for a user setting object of the specified
-    /// <paramref name="settingType"/>.
+    /// Gets the SettingAttribute of a property.
     /// </summary>
-    /// <param name="settingType">Type of the settings class to map to a filename.</param>
-    /// <returns>File name without path of a file which will store the setting instance of the
-    /// specified <paramref name="settingType"/>.</returns>
-    protected static string GetUserFilePath(Type settingType)
+    /// <param name="property">Property to assess.</param>
+    /// <returns>Propertie's meta attribute or <c>null</c>, if no <see cref="SettingAttribute"/> was
+    /// attached to the specified <paramref name="property"/>.</returns>
+    protected static SettingAttribute GetSettingAttribute(ICustomAttributeProvider property)
     {
-      string fullUserFileName = String.Format(@"<CONFIG>\{0}\{1}", Environment.UserName, settingType.FullName + ".xml");
+      // Get the info stored in the SettingAttribute, if available
+      object[] attributes = property.GetCustomAttributes(typeof(SettingAttribute), false);
+      if (attributes.Length != 0)
+        return (SettingAttribute)attributes[0];
+      return null;
+    }
+
+    /// <summary>
+    /// Returns the full file path for a user setting object of the specified
+    /// <paramref name="settingsType"/>.
+    /// </summary>
+    /// <param name="settingsType">Type of the settings class to map to a filename.</param>
+    /// <returns>File name without path of a file which will store the setting instance of the
+    /// specified <paramref name="settingsType"/>.</returns>
+    protected static string GetUserFilePath(Type settingsType)
+    {
+      string fullUserFileName = String.Format(@"<CONFIG>\{0}\{1}", Environment.UserName, settingsType.FullName + ".xml");
       return ServiceScope.Get<IPathManager>().GetPath(fullUserFileName);
     }
 
     /// <summary>
     /// Returns the full file path for a global setting object of the specified
-    /// <paramref name="settingType"/>.
+    /// <paramref name="settingsType"/>.
     /// </summary>
-    /// <param name="settingType">Type of the settings class to map to a filename.</param>
+    /// <param name="settingsType">Type of the settings class to map to a filename.</param>
     /// <returns>File name without path of a file which will store the setting instance of the
-    /// specified <paramref name="settingType"/>.</returns>
-    protected static string GetGlobalFilePath(Type settingType)
+    /// specified <paramref name="settingsType"/>.</returns>
+    protected static string GetGlobalFilePath(Type settingsType)
     {
-      string fullFileName = String.Format(@"<CONFIG>\{0}", settingType.FullName + ".xml");
+      string fullFileName = String.Format(@"<CONFIG>\{0}", settingsType.FullName + ".xml");
       return ServiceScope.Get<IPathManager>().GetPath(fullFileName);
     }
 
     #endregion
+
+    protected object LoadSettingsObject(Type settingsType)
+    {
+      SettingsFileHandler globalHandler = new SettingsFileHandler(GetGlobalFilePath(settingsType));
+      SettingsFileHandler userHandler = new SettingsFileHandler(GetUserFilePath(settingsType));
+      try
+      {
+        globalHandler.Load();
+      }
+      catch (Exception e)
+      {
+        ServiceScope.Get<ILogger>().Error("SettingsManager: Error loading global settings file for setting type '{0}'... Will clear this settings file.", e, settingsType.Name);
+        globalHandler.Clear();
+        RemoveSettingsData(settingsType, false, true);
+      }
+      try
+      {
+        userHandler.Load();
+      }
+      catch (Exception e)
+      {
+        ServiceScope.Get<ILogger>().Error("SettingsManager: Error loading user settings file for setting type '{0}'... Will clear this settings file.", e, settingsType.Name);
+        userHandler.Clear();
+        RemoveSettingsData(settingsType, true, false);
+      }
+      try
+      {
+        object result = Activator.CreateInstance(settingsType);
+        foreach (PropertyInfo property in result.GetType().GetProperties())
+        {
+          SettingAttribute att = GetSettingAttribute(property);
+          if (att == null) continue;
+          SettingsFileHandler s = (att.SettingScope == SettingScope.Global ? globalHandler : userHandler);
+          object value = s.GetValue(property.Name, property.PropertyType);
+          if (value == null)
+            if (att.SetDefault)
+              value = att.DefaultValue;
+            else
+              continue;
+          property.SetValue(result, value, null);
+        }
+        return result;
+      }
+      catch (Exception e)
+      {
+        ServiceScope.Get<ILogger>().Error("SettingsManager: Error loading settings of type '{0}'", e, settingsType.Name);
+        return null;
+      }
+    }
+
+    protected void SaveSettingsObject(object settingsObject)
+    {
+      try
+      {
+        Type t = settingsObject.GetType();
+        SettingsFileHandler globalSerializer = new SettingsFileHandler(GetGlobalFilePath(t));
+        SettingsFileHandler userSerializer = new SettingsFileHandler(GetUserFilePath(t));
+        foreach (PropertyInfo property in t.GetProperties())
+        {
+          SettingAttribute att = GetSettingAttribute(property);
+          if (att == null) continue;
+          SettingsFileHandler s = (att.SettingScope == SettingScope.Global ? globalSerializer : userSerializer);
+          object value = property.GetValue(settingsObject, null);
+          s.SetValue(property.Name, value);
+        }
+        globalSerializer.Close();
+        userSerializer.Close();
+      }
+      catch (Exception e)
+      {
+        ServiceScope.Get<ILogger>().Error("SettingsManager: Error writing settings of type '{0}'... Will clear settings files for this setting", e, settingsObject.GetType().Name);
+        RemoveSettingsData(settingsObject.GetType(), true, true);
+      }
+    }
 
     #region ISettingsManager implementation
 
@@ -94,8 +185,7 @@ namespace MediaPortal.Core.Services.Settings
       object result;
       if ((result = _cache.Get(settingsType)) != null)
         return result;
-      SettingParser parser = new SettingParser(GetGlobalFilePath(settingsType), GetUserFilePath(settingsType));
-      result = parser.Deserialize(settingsType);
+      result = LoadSettingsObject(settingsType);
       _cache.Set(result);
       return result;
     }
@@ -107,10 +197,7 @@ namespace MediaPortal.Core.Services.Settings
       _cache.Set(settingsObject);
       if (_batchUpdate)
         return;
-      SettingParser parser = new SettingParser(
-          GetGlobalFilePath(settingsObject.GetType()),
-          GetUserFilePath(settingsObject.GetType()));
-      parser.Serialize(settingsObject);
+      SaveSettingsObject(settingsObject);
     }
 
     public void StartBatchUpdate()
@@ -128,19 +215,39 @@ namespace MediaPortal.Core.Services.Settings
       _cache.StopKeep();
     }
 
+    public void RemoveSettingsData(Type settingsType, bool user, bool global)
+    {
+      if (user)
+      {
+        string userPath = GetUserFilePath(settingsType);
+        FileInfo userConfigFile = new FileInfo(userPath);
+        if (userConfigFile.Exists)
+          userConfigFile.Delete();
+      }
+      if (global)
+      {
+        string globalPath = GetGlobalFilePath(settingsType);
+        FileInfo globalConfigFile = new FileInfo(globalPath);
+        if (globalConfigFile.Exists)
+          globalConfigFile.Delete();
+      }
+    }
+
     public void RemoveAllSettingsData(bool user, bool global)
     {
       if (user)
       {
         string userPath = ServiceScope.Get<IPathManager>().GetPath(string.Format(@"<CONFIG>\{0}", Environment.UserName));
         DirectoryInfo userConfigDirectory = new DirectoryInfo(userPath);
-        userConfigDirectory.Delete(true);
+        if (userConfigDirectory.Exists)
+          userConfigDirectory.Delete(true);
       }
       if (global)
       {
         string globalPath = ServiceScope.Get<IPathManager>().GetPath("<CONFIG>");
         DirectoryInfo globalConfigDirectory = new DirectoryInfo(globalPath);
-        globalConfigDirectory.Delete(true);
+        if (globalConfigDirectory.Exists)
+          globalConfigDirectory.Delete(true);
       }
     }
